started work breaking out the bridge into multiple files
