--- pytest.ini
@@ -1,32 +1,0 @@
-[pytest]
-minversion = 6.0
-addopts = 
-    -ra
-    -q
-    --strict-markers
-    --strict-config
-    --tb=short
-    --cov=bridge_mcp_ghidra
-    --cov=tools
-    --cov=debugger
-    --cov-report=term-missing:skip-covered
-    --cov-report=html:tests/coverage
-    --cov-report=xml:tests/coverage.xml
-    --junit-xml=tests/junit.xml
-testpaths = tests
-python_files = test_*.py
-python_classes = Test*
-python_functions = test_*
-markers =
-    slow: marks tests as slow (deselect with '-m "not slow"')
-    integration: marks tests as integration tests
-    functional: marks tests as functional tests
-    unit: marks tests as unit tests
-    api: marks tests as API endpoint tests
-    requires_ghidra: marks tests that require Ghidra server to be running
-    requires_binary: marks tests that require a loaded binary in Ghidra
-    readonly: marks tests that only read data and do not modify Ghidra state
-    safe_write: marks tests that write data but use identity operations (write same value back)
-filterwarnings =
-    ignore::DeprecationWarning
-    ignore::PendingDeprecationWarning
