+++ python/bridge_mcp_ghidra/__main__.py
@@ -1,0 +1,6 @@
+"""Enable ``python -m bridge_mcp_ghidra``."""
+
+from .cli import main
+
+if __name__ == "__main__":
+    main()
