--- requirements-test.txt
@@ -1,26 +1,0 @@
-# Test Requirements for GhidraMCP Test Suite
-# Install with: pip install -r requirements-test.txt
-
-# Core testing framework
-pytest>=7.0.0
-pytest-cov>=7.1.0
-pytest-xdist>=3.8.0
-pytest-timeout>=2.4.0
-pytest-mock>=3.10.0
-pytest-asyncio>=1.4.0
-
-# HTTP testing
-requests>=2.32.0
-responses>=0.26.1
-requests-mock>=1.12.1
-
-# Test reporting
-coverage>=7.0.0
-pytest-json-report>=1.5.0
-
-# YAML parsing (workflow validation tests)
-pyyaml>=6.0.0
-
-# Benchmark truth extraction (libclang-backed C parser)
-clang>=18.1.8
-libclang>=18.1.1
