--- requirements.txt
@@ -1,20 +1,0 @@
-# GhidraMCP - Production Dependencies
-# Install with: pip install -r requirements.txt
-
-# Core MCP server framework
-mcp>=1.27.2,<2.0.0
-
-# HTTP client library
-# 2.32.0 is the floor for CVE-2024-35195 (certificate verification bypass).
-requests>=2.32.0,<3.0.0
-
-# URL parsing utilities (included in stdlib but explicit for clarity)
-# urllib.parse - included in Python standard library
-
-# JSON handling (included in stdlib)
-# json - included in Python standard library
-
-# For development and testing, install additional requirements:
-# pip install -r requirements-test.txt
-# For the standalone debugger server on Windows, also install:
-# pip install -r requirements-debugger.txt
